Refactor router code
