+++ src/microunit/src/node_server.rs
@@ -14,9 +14,9 @@
 
 use std::{net::SocketAddr, sync::Arc};
 
-use axum::{Router, Server};
+use axum::Server;
 
-use super::{error::Result, node::Node};
+use crate::{error::Result, node::Node, node_router::route};
 
 /// An HTTP server that serves a node.
 pub struct NodeServer {
@@ -31,60 +31,10 @@ impl NodeServer {
     }
 
     pub async fn bind(&self, addr: SocketAddr) -> Result<()> {
-        let v1 = v1::route(self.node.clone());
-        let router = Router::new().nest("/v1", v1);
+        let router = route(self.node.clone());
         Server::bind(&addr)
             .serve(router.into_make_service())
             .await?;
         Ok(())
-    }
-}
-
-mod v1 {
-    use std::sync::Arc;
-
-    use axum::{
-        extract::Extension, http::StatusCode, response::IntoResponse, routing::get,
-        AddExtensionLayer, Json, Router,
-    };
-    use serde_json::json;
-
-    use crate::{node::Node, unit::UnitSpec};
-
-    pub fn route(node: Arc<Node>) -> Router {
-        Router::new()
-            .route("/status", get(status))
-            .route("/units", get(list_units).post(create_unit))
-            .layer(AddExtensionLayer::new(node))
-    }
-
-    async fn status(Extension(node): Extension<Arc<Node>>) -> impl IntoResponse {
-        let desc = node.status();
-        (StatusCode::OK, Json(desc))
-    }
-
-    async fn list_units(Extension(node): Extension<Arc<Node>>) -> impl IntoResponse {
-        let descs = node.list_units().await;
-        (StatusCode::OK, Json(descs))
-    }
-
-    async fn create_unit(
-        Json(spec): Json<UnitSpec>,
-        Extension(node): Extension<Arc<Node>>,
-    ) -> impl IntoResponse {
-        match node.create_unit(spec).await {
-            Ok(desc) => {
-                let resp = json!({
-                    "desc": desc,
-                });
-                (StatusCode::CREATED, Json(resp))
-            }
-            Err(err) => {
-                let resp = json!({
-                    "error": err.to_string(),
-                });
-                (StatusCode::INTERNAL_SERVER_ERROR, Json(resp))
-            }
-        }
     }
 }
